Add details to README and add docx & pptx modules to requirements
The extra readme details contain basic stuff like where to download different models.
Or how it's a good idea to upgrade pip, or use a virtual environment.
Also added a link to the tutorial with screenshots for ease of adoption.

Finally, python-docx and python-pptx were missing in requirements.txt.
Ingesting word and powerpoint documents now work.