Add phi-2 support
This patch introduces the chatml prompt format (and associated tests) necessary
for driving the Dolphin 2.6 fine-tune of microsoft/phi-2.  I'm using the
fine-tune rather because the prompt behaviour with the chatml format seems to
be more robust than the upstream.

Also included is `settings-phi-2.yaml`, which you can use by setting
`PGPT_PROFILES=phi-2` for a local experience that's tolerable on CPU.  On an M1
mac with Metal, I'm getting about 20 tokens per second output using the Q4_K_M
model listed.

The phi-2 architecture also needs the `llama-cpp-python` version to be updated,
so a bump to 0.2.28 is included.